AP-12791: Changed prop handling and testing
Removed default props and replaced with custom validators for the Double and Integer inputs. Updated the tests for the NumberInput, DoubleInput and IntegerInput components. There are some Vue warnings being created by the vue-svg-loader, but these are non-critical and will be fixed in the future.

AP-12791 (Frontend: rewrite integer input widget)