AP-12903: Added boolean input widget
Created BooleanInput component, registered it with the widgets.config.js and added it to the Widget component registry. Also modified the Checkbox component in webapp commons to make it more reusable. Changed a util function to correctly process values that are not of type string or number. Lastly, changed default props on the Text Output widget to use validators instead of empty object.

AP-12903 (Frontend: rewrite boolean input widgets)