WEBP-168: Widget.vue: wait for update value to be finished before calling is valid.
Otherwise Widgets (and their sub components) can not use their value prop to check if it is valid or not and need to manage the state again. We did this a lot and should stop doing it!
Bug was found in Dropdown.vue.

WEBP-168 (Frontend: rewrite nominal row filter selection widget (selection))