UIEXT-227: Disable 'execute'-button if node can't be executed
... in 'ViewExecutable'-overlay. Also display a respective text that the
node can't be executed.
A node is currently regarded to be non-executable if it's in 'idle'
state. Even though nodes could be executed if in idle state, we don't
allow it for now because otherwise we would need a mechanism to update
the input specs for an opened dialog (which we don't have).
Test added, too.

UIEXT-227 (Nodes should be able to open dialog even if no data is connected)