UIEXT-677: Page-builder build-artefact updated
The page-builder artefacts are registered at the (legacy)
web-resource extension point. Both aspects of the extension point
registration updated: the reference to the actual artefact-file as well
as the relative path at which it's exposed to the frontend.

UIEXT-677 (Use static pagebuilder html page (AP) and update all pagebuilder-resource references)